refactor: split onboard flow support
